Move generator to a subfolder to allow for easy addition of new languages
