+++ pkg/generator/golang/generator.go
@@ -11,12 +11,12 @@
 	limitations under the License.
 */
 
-package generator
+package golang
 
 import (
 	"bytes"
 	"errors"
-	"github.com/loopholelabs/scale-signature/pkg/generator/templates"
+	"github.com/loopholelabs/scale-signature/pkg/generator/golang/templates"
 	"github.com/loopholelabs/scale-signature/pkg/schema"
 	"go/format"
 	"text/template"
