+++ pkg/generator/golang/templates/go_arrays.templ
