+++ pkg/generator/golang/templates/go_enumarrays.templ
