+++ pkg/generator/golang/templates/go_enummaps.templ
