+++ pkg/generator/golang/templates/go_enums.templ
