+++ pkg/generator/golang/templates/go_maps.templ
