+++ pkg/generator/golang/templates/go_modelarrays.templ
