+++ pkg/generator/golang/templates/go_modelmaps.templ
