+++ pkg/generator/golang/templates/go_models.templ
