+++ pkg/generator/golang/templates/go_primitives.templ
