+++ pkg/generator/golang/templates/templates.go
