+++ pkg/generator/golang/templates/types.go.templ
